Extract IO-oriented tests
